Изменён интерфейс для возможности наследования локальным и удалённым сервисом. Изменена обработка ошибок. Удалён неиспольуемый метод. Мелкий рефакторинг.
